[ASV-1197] Removed unused code; Extracted methods for better readability; Extracted CampaignsService response to it's own class to better support extending the service; Small renames and bug fixes;
