+++ app/src/main/java/cm/aptoide/pt/ads/CampaignsServiceResponse.java
@@ -7,13 +7,19 @@ public class CampaignsServiceResponse {
   private String next;
   private List<Campaign> campaigns;
 
-  public CampaignsServiceResponse(String next, List<Campaign> campaigns, boolean hasError) {
+  public CampaignsServiceResponse(String next, List<Campaign> campaigns) {
     this.next = next;
     this.campaigns = campaigns;
-    this.hasError = hasError;
+    this.hasError = false;
   }
 
-  public boolean isHasError() {
+  public CampaignsServiceResponse() {
+    this.hasError = true;
+    this.next = null;
+    this.campaigns = null;
+  }
+
+  public boolean hasError() {
     return hasError;
   }
 
@@ -32,22 +38,19 @@ public class CampaignsServiceResponse {
     private int downloads;
     private String packageName;
     private float averageRating;
-    private long totalRatings;
     private String appc;
     private String clickUrl;
     private String downloadUrl;
     private String versionName;
 
     public Campaign(String uid, String label, String icon, int downloads, String packageName,
-        float averageRating, long totalRatings, String appc, String clickUrl, String downloadUrl,
-        String versionName) {
+        float averageRating, String appc, String clickUrl, String downloadUrl, String versionName) {
       this.uid = uid;
       this.label = label;
       this.icon = icon;
       this.downloads = downloads;
       this.packageName = packageName;
       this.averageRating = averageRating;
-      this.totalRatings = totalRatings;
       this.appc = appc;
       this.clickUrl = clickUrl;
       this.downloadUrl = downloadUrl;
@@ -72,10 +75,6 @@ public class CampaignsServiceResponse {
 
     public float getAverageRating() {
       return averageRating;
-    }
-
-    public long getTotalRatings() {
-      return totalRatings;
     }
 
     public String getAppc() {
