+++ app/src/main/java/cm/aptoide/pt/app/view/MoreBundleView.java
@@ -9,6 +9,4 @@ import cm.aptoide.pt.home.apps.BundleView;
 public interface MoreBundleView extends BundleView {
 
   void setToolbarInfo(String title);
-
-  void showRewardAppSnack();
 }
