+++ app/src/main/java/cm/aptoide/pt/home/HomeView.java
@@ -48,6 +48,4 @@ public interface HomeView extends BundleView {
   void setEllipsizedPromotionsTicker();
 
   void hidePromotionsIcon();
-
-  void showRewardAppSnack();
 }
