+++ gradle.properties
@@ -30,12 +30,12 @@ APTOIDE_WEB_SERVICES_V7_HOST_DEVELOPMENT=dev.ws75.aptoide.com
 APTOIDE_WEB_SERVICES_READ_V7_HOST_DEVELOPMENT=dev.ws75-secondary.aptoide.com
 APTOIDE_WEB_SERVICES_WRITE_V7_HOST_DEVELOPMENT=dev.ws75-primary.aptoide.com
 APTOIDE_WEB_SERVICES_APTWORDS_HOST_DEVELOPMENT=dev.aptoide.com
+APTOIDE_WEB_SERVICES_SCHEME_DEVELOPMENT=http
 #aptoide hostnames
 APTOIDE_WEB_SERVICES_SEARCH_HOST=buzz.aptoide.com:10001
 APTOIDE_WEB_SERVICES_AB_TESTING_HOST=abtesting.aptoide.com
 APTOIDE_WEB_SERVICES_APICHAIN_BDS_HOST=apichain.blockchainds.com
 APTOIDE_WEB_SERVICES_API_BDS_HOST=api.blockchainds.com
-APTOIDE_WEB_SERVICES_SCHEME_DEVELOPMENT=http
 APTOIDE_WEBSOCKETS_HOST=buzz.webservices.aptoide.com
 APTOIDE_WEBSOCKETS_SCHEME=ws://
 APTOIDE_WEB_SERVICES_NOTIFICATION_HOST=pnp.aptoide.com
